chore: rename addressbookscreen to v1
Update test suite name to not use v2 naming
