+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/screens/TokensVsUtxoFaq.test.tsx
@@ -3,7 +3,7 @@ import { TokensVsUtxoFaq } from './TokensVsUtxoFaq'
 
 jest.mock('@shared-contexts/ThemeProvider')
 
-describe('Tokens vs Utxo FAQ V2 screen', () => {
+describe('Tokens vs Utxo FAQ screen', () => {
   it('should match snapshot', async () => {
     const rendered = render(<TokensVsUtxoFaq />)
     expect(rendered.toJSON()).toMatchSnapshot()
