+++ mobile-app/app/screens/AppNavigator/screens/Settings/SettingsNavigator.tsx
@@ -21,7 +21,7 @@ import { TokensVsUtxoFaq } from '../Portfolio/screens/TokensVsUtxoFaq'
 import { ChangePinScreen } from '@screens/AppNavigator/screens/Settings/screens/ChangePinScreen'
 import { ConfirmPinScreen } from '@screens/AppNavigator/screens/Settings/screens/ConfirmPinScreen'
 import { RecoveryWordsScreen } from './screens/RecoveryWordsScreen'
-import { AddressBookScreen } from '../Portfolio/screens/AddressBookScreenV2'
+import { AddressBookScreen } from '../Portfolio/screens/AddressBookScreen'
 import { LocalAddress } from '@store/userPreferences'
 import { AddOrEditAddressBookScreen } from '../Portfolio/screens/AddOrEditAddressBookScreen'
 import { NetworkSelectionScreen } from '@screens/AppNavigator/screens/Settings/screens/NetworkSelectionScreen'
