+++ mobile-app/app/screens/WalletNavigator/screens/CreateWallet/DexFaq.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react-native'
 import { DexFaq } from './DexFaq'
 
 jest.mock('@shared-contexts/ThemeProvider')
-describe('DEX FAQ V2 screen', () => {
+describe('DEX FAQ screen', () => {
   it('should match snapshot', async () => {
     const rendered = render(<DexFaq />)
     expect(rendered.toJSON()).toMatchSnapshot()
