+++ mobile-app/app/screens/WalletNavigator/screens/CreateWallet/LiquidityMiningFaq.test.tsx
@@ -2,7 +2,7 @@ import { render } from '@testing-library/react-native'
 import { LiquidityMiningFaq } from './LiquidityMiningFaq'
 
 jest.mock('@shared-contexts/ThemeProvider')
-describe('Liquidity Mining V2 FAQ screen', () => {
+describe('Liquidity Mining FAQ screen', () => {
   it('should match snapshot', async () => {
     const rendered = render(<LiquidityMiningFaq />)
     expect(rendered.toJSON()).toMatchSnapshot()
